v0.69 PR2 (Phase 2 complete) — split remaining 10 generate.ts subcommands
Completes Plan G Phase 2 (generate.ts per-subcommand split). Combined with
PR1 (#164), all 13 subcommands now live in their own files under
`packages/cli/src/commands/generate/`. `generate.ts` collapses to a
165-line thin barrel: imports each `register*Command` and calls it on the
top-level `generateCommand`.

## Splits in this PR (10)

| File | Lines | Source |
|---|---:|---|
| `generate/image.ts` | 537 | OpenAI/Gemini/Grok/Runway image |
| `generate/video.ts` | 569 | fal/Grok/Veo/Kling/Runway video |
| `generate/speech.ts` | 237 | ElevenLabs TTS + duration-fit |
| `generate/music.ts` | 310 | ElevenLabs Music + Replicate MusicGen |
| `generate/storyboard.ts` | 211 | Claude script-to-storyboard |
| `generate/thumbnail.ts` | 192 | DALL-E + Gemini best-frame |
| `generate/background.ts` | 193 | OpenAI gpt-image-2 backdrops |
| `generate/video-status.ts` | 295 | hidden — Grok/Runway/Kling status |
| `generate/video-extend.ts` | 276 | hidden — Kling/Veo extend |
| (`generate/video-cancel.ts`, `music-status.ts`, `sound-effect.ts` already in PR1) |  |  |

`generate.ts`: 2,324 → 165 lines (−93%).

## Pattern

Each subcommand file:
1. Library export: `executeXxx` + `ExecuteXxxOptions/Result` (when used by
   pipeline executor or manifest)
2. CLI export: `registerXxxCommand(parent: Command): void`

`generate.ts` re-exports library functions for backward compat (callers
in pipeline/executor.ts and manifest/generate.ts continue to import from
`./generate.js`). Provider-resolver inline maps, getStatusColor helper,
all extracted execute functions all moved to per-subcommand files.

## Behavior preservation

- 652 cli tests + 46 mcp tests + lint 0 errors all green
- `vibe generate --help` lists same 9 visible commands (4 hidden)
- `vibe generate image --help` byte-identical
- All `executeXxx` re-exports work for pipeline/executor.ts

## Out of scope (future)

- Manifest path updates: `manifest/generate.ts` still imports through
  `generate.ts` re-export shim — works, but a follow-up could update to
  direct `from "./generate/<sub>.js"` imports
- v0.69.0 version bump waits until any other Phase 2/3 cleanup lands

Net: +3,177 lines new files / −2,194 deletions = +983 lines (mostly
imports + boilerplate per file). generate.ts itself shrank 2,159 lines.
